Time: 0 ms (100.00%) | Memory: 8.1 MB (12.18%) - LeetSync

+++ 1448-maximum-69-number/maximum-69-number.cpp
@@ -7,7 +7,7 @@ public:
         temp /= 10;
         cnt++;
     }
-       return (pos==-1)?num: int(num + (3*pow(10,pos)));
+       return (pos==-1)?num: (num + (3*pow(10,pos)));
     }
 };
